fix: project caching
