+++ src/app.ts
@@ -1,6 +1,7 @@
 // Bitgreen API Server
 
 /* import packages */
+import cron from 'node-cron';
 import cookieParser from 'cookie-parser';
 import cors, { CorsOptions } from 'cors';
 import * as dotenv from 'dotenv';
@@ -17,6 +18,7 @@ import cart from './routes/ccMarketplace/profile/cart';
 import profile from './routes/ccMarketplace/profile/profile';
 import creditTransaction from './routes/ccMarketplace/dex/credit-transaction';
 import transaction from './routes/transaction';
+import {updateProjectsCron} from "./services/update-projects";
 
 /* config */
 dotenv.config();
@@ -71,9 +73,9 @@ const mainLoop = async () => {
   const server = app.listen(port, function () {
     console.log(`API server is listening at: http://localhost:${port}.`);
   });
+
+  await updateProjectsCron()
 };
 
 // run main function
 mainLoop().catch(console.error);
-
-// write a fetch command in javascript
