+++ src/services/methods/blockExtrinsic.ts
@@ -5,6 +5,7 @@ import {
   EventRecord,
 } from '@polkadot/types/interfaces';
 import { Vec } from '@polkadot/types-codec';
+import {SignedBlock} from "@polkadot/types/interfaces/runtime/types";
 
 export async function blockExtrinsic(
   api: ApiPromise,
@@ -20,8 +21,19 @@ export async function blockExtrinsic(
       api.rpc.chain.getBlock(blockHash),
       apiAt.query.system.events() as Promise<Vec<EventRecord>>,
     ]);
-    return { signedBlock: signedBlock, blockEvents: blockEvents };
+
+    const blockDate = await getBlockDate(api, signedBlock)
+
+    return { signedBlock: signedBlock, blockEvents: blockEvents, blockDate };
   } catch (error) {
     return {};
   }
+}
+
+export async function getBlockDate(api: ApiPromise, signedBlock: SignedBlock) {
+  const blockTimestamp = signedBlock!.block.extrinsics.find(
+      (extrinsic) => extrinsic.method.section === 'timestamp' && extrinsic.method.method === 'set'
+  );
+
+  return new Date(Number(blockTimestamp!.method.args[0].toString()))
 }
